refactor: wave13 peels — planning, dev-server, interviews, task-artifacts-ops
Behavior-preserving package organization after wave 12:
- app/api: planning, dev-server, mission-interview, models-usage,
  task-steer, ai-summarize (legacy re-exports kept)
- rename remaining-ops-7 → task-store/task-artifacts-ops.ts
- legacy.ts shrinks below the 2000-line cap (~3342 → ~1408)
