+++ packages/dashboard/app/api/ai-summarize.ts
@@ -1,0 +1,67 @@
+/**
+ * FNXC:CodeOrganization 2026-07-20-14:00:
+ * AI title summarization client API peeled from legacy.ts.
+ */
+
+import { withTokenHeader } from "../auth";
+
+// --- AI Summarization API ---
+
+/** Response from title summarization endpoint */
+export interface SummarizeTitleResponse {
+  title: string;
+}
+
+/** Summarize a task description into a concise title using AI.
+ * @param description - The task description to summarize (must be >200 chars; model input is truncated)
+ * @param provider - Optional AI model provider (e.g., "anthropic")
+ * @param modelId - Optional AI model ID (e.g., "claude-sonnet-4-5")
+ * @param projectId - Optional project ID for scoped settings resolution
+ * @returns The generated title (guaranteed ≤60 characters)
+ * @throws Error with descriptive message for 400/429/503 errors
+ */
+export async function summarizeTitle(
+  description: string,
+  provider?: string,
+  modelId?: string,
+  projectId?: string
+): Promise<string> {
+  const url = projectId
+    ? `/api/ai/summarize-title?projectId=${encodeURIComponent(projectId)}`
+    : "/api/ai/summarize-title";
+  const res = await fetch(url, {
+    method: "POST",
+    headers: withTokenHeader({ "Content-Type": "application/json" }),
+    body: JSON.stringify({ description, provider, modelId }),
+  });
+
+  const contentType = res.headers.get("content-type") ?? "";
+  const bodyText = await res.text();
+  const isJson = contentType.includes("application/json");
+
+  if (!isJson) {
+    throw new Error(`API returned non-JSON response: ${bodyText.slice(0, 100)}`);
+  }
+
+  const data = JSON.parse(bodyText) as { title?: string; error?: string };
+
+  if (!res.ok) {
+    const errorMessage = data.error || "Request failed";
+    if (res.status === 400) {
+      throw new Error(`Invalid request: ${errorMessage}`);
+    } else if (res.status === 429) {
+      throw new Error(`Rate limit exceeded: ${errorMessage}`);
+    } else if (res.status === 503) {
+      throw new Error(`AI service temporarily unavailable: ${errorMessage}`);
+    } else {
+      throw new Error(errorMessage);
+    }
+  }
+
+  if (!data.title) {
+    throw new Error("API returned empty title");
+  }
+
+  return data.title;
+}
+
