+++ packages/dashboard/app/api/legacy.ts
@@ -1,17 +1,8 @@
-import type { SubtaskItem, PlanningSubtaskDraft } from "./ai-text.js";
-import type { AgentCapability } from "@fusion/core";
 import type {
-  Task,
   CommitAssociationDiffBackfillReport,
 } from "@fusion/core";
 // Consumers import backfill report types from the legacy API barrel.
 export type { CommitAssociationDiffBackfillReport };
-import type {
-  PlanningQuestion,
-  PlanningSummary,
-} from "@fusion/core";
-import type { MissionInterviewDraftSummary } from "../components/mission-types";
-import { withTokenHeader } from "../auth";
 
 /* FNXC:DashboardApi 2026-07-15-13:25: Preserve the legacy API barrel while consumers migrate to focused modules. */
 export {
@@ -275,11 +266,8 @@ export type {
   QmdInstallResult,
 } from "./memory.js";
 
-import { api, buildApiUrl } from "./client.js";
-import { withProjectId } from "./health.js";
 
-// Import + re-export skills types so legacy monofile bodies can reference them
-// while hooks/components keep stable import paths via this barrel.
+// Re-export skills types so hooks/components keep stable import paths via this barrel.
 import type {
   DiscoveredSkill,
   CatalogEntry,
@@ -299,97 +287,32 @@ export type {
   SkillFileContent,
 };
 
-export function addSteeringComment(id: string, text: string, projectId?: string): Promise<Task> {
-  return api<Task>(withProjectId(`/tasks/${id}/steer`, projectId), {
-    method: "POST",
-    body: JSON.stringify({ text }),
-  });
-}
+/*
+ * FNXC:CodeOrganization 2026-07-20-14:00:
+ * Preserve legacy `task-steer` imports while implementations live in task-steer.js.
+ */
+export {
+  addSteeringComment,
+  requestSpecRevision,
+  rebuildTaskSpec,
+  refineTask,
+} from "./task-steer.js";
 
-export function requestSpecRevision(id: string, feedback: string, projectId?: string): Promise<Task> {
-  return api<Task>(withProjectId(`/tasks/${id}/spec/revise`, projectId), {
-    method: "POST",
-    body: JSON.stringify({ feedback }),
-  });
-}
-
-export function rebuildTaskSpec(id: string, projectId?: string): Promise<Task> {
-  return api<Task>(withProjectId(`/tasks/${id}/spec/rebuild`, projectId), {
-    method: "POST",
-  });
-}
-
-export function refineTask(id: string, feedback: string, projectId?: string): Promise<Task> {
-  return api<Task>(withProjectId(`/tasks/${id}/refine`, projectId), {
-    method: "POST",
-    body: JSON.stringify({ feedback }),
-  });
-}
-
-
-// --- Models API ---
-
-/** Available AI model info returned by the models endpoint */
-export interface ModelInfo {
-  provider: string;
-  id: string;
-  name: string;
-  reasoning: boolean;
-  contextWindow: number;
-}
-
-/** Response from the models endpoint */
-export interface ModelsResponse {
-  models: ModelInfo[];
-  favoriteProviders: string[];
-  favoriteModels: string[];
-  defaultProvider?: string;
-  defaultModelId?: string;
-  resolvedPlanningProvider?: string;
-  resolvedPlanningModelId?: string;
-}
-
-/** Fetch available AI models from the model registry along with favoriteProviders */
-export function fetchModels(): Promise<ModelsResponse> {
-  return api<ModelsResponse>("/models");
-}
-
-// --- Usage API ---
-
-/** Pace information for weekly usage windows */
-export interface UsagePace {
-  status: "ahead" | "on-track" | "behind";
-  percentElapsed: number; // 0-100, how much of the window time has passed
-  message: string; // e.g., "Using 15% over your limit pace"
-}
-
-/** Usage window for a provider (e.g., "Session (5h)", "Weekly") */
-export interface UsageWindow {
-  label: string;
-  percentUsed: number; // 0-100
-  percentLeft: number; // 0-100
-  resetText: string | null; // e.g., "resets in 2h"
-  resetMs?: number; // ms until reset
-  resetAt?: string; // ISO 8601 timestamp of when the window resets (machine-readable)
-  windowDurationMs?: number; // total window length
-  pace?: UsagePace; // pace indicator for weekly windows
-}
-
-/** Provider usage data */
-export interface ProviderUsage {
-  name: string;
-  icon: string; // emoji
-  status: "ok" | "error" | "no-auth";
-  error?: string;
-  plan?: string | null;
-  email?: string | null;
-  windows: UsageWindow[];
-}
-
-/** Fetch usage data from all configured AI providers */
-export function fetchUsageData(): Promise<{ providers: ProviderUsage[] }> {
-  return api<{ providers: ProviderUsage[] }>("/usage");
-}
+/*
+ * FNXC:CodeOrganization 2026-07-20-14:00:
+ * Preserve legacy `models-usage` imports while implementations live in models-usage.js.
+ */
+export {
+  fetchModels,
+  fetchUsageData,
+} from "./models-usage.js";
+export type {
+  ModelInfo,
+  ModelsResponse,
+  UsagePace,
+  UsageWindow,
+  ProviderUsage,
+} from "./models-usage.js";
 
 /*
  * FNXC:CodeOrganization 2026-07-20-10:00:
@@ -659,1360 +582,96 @@ export type {
   WorkspaceTaskInfo,
 } from "./workspace-files.js";
 
-// --- Planning Mode API ---
-
-/** Planning session state returned from API */
-export interface PlanningSession {
-  sessionId: string;
-  currentQuestion: PlanningQuestion | null;
-  summary: PlanningSummary | null;
-}
-
-/** The response endpoint may synchronously return a generated next question before SSE delivers it. */
-export type PlanningResponse = PlanningSession | { type: "question"; data: PlanningQuestion };
-
-/** SSE event types for planning session streaming */
-export type PlanningStreamEvent =
-  | { type: "thinking"; data: string }
-  | { type: "question"; data: PlanningQuestion }
-  | { type: "summary"; data: PlanningSummary }
-  | { type: "error"; data: string }
-  | { type: "complete"; data: Record<string, never> };
-
-export interface AgentOnboardingSummary {
-  name: string;
-  role: AgentCapability | "custom";
-  instructionsText: string;
-  thinkingLevel: "off" | "minimal" | "low" | "medium" | "high" | "xhigh";
-  maxTurns: number;
-  title?: string;
-  icon?: string;
-  reportsTo?: string;
-  soul?: string;
-  memory?: string;
-  skills?: string[];
-  templateId?: string;
-  patternAgentId?: string;
-  rationale?: string;
-  model?: string;
-  /** Draft-only AI suggestion for eventual runtimeConfig.model selection. */
-  modelHint?: string;
-  /** Draft-only AI suggestion for eventual runtimeConfig.runtimeHint plugin runtime selection. */
-  runtimeHint?: string;
-  heartbeatProcedurePath?: string;
-  heartbeatIntervalMs?: number;
-  heartbeatEnabled?: boolean;
-}
-
-export type OnboardingMode = "create" | "edit";
-export type ThinkingLevel = "off" | "minimal" | "low" | "medium" | "high" | "xhigh";
-
-export interface ExistingAgentOnboardingConfig {
-  name?: string;
-  role?: AgentCapability | "custom";
-  title?: string;
-  instructionsText?: string;
-  soul?: string;
-  memory?: string;
-  reportsTo?: string;
-  skills?: string[];
-  model?: string;
-  thinkingLevel?: ThinkingLevel;
-  maxTurns?: number;
-  runtimeHint?: string;
-  heartbeatIntervalMs?: number;
-  heartbeatTimeoutMs?: number;
-  maxConcurrentRuns?: number;
-  messageResponseMode?: "immediate" | "on-heartbeat";
-}
-
-export type AgentOnboardingStreamEvent =
-  | { type: "thinking"; data: string }
-  | { type: "question"; data: PlanningQuestion }
-  | { type: "summary"; data: AgentOnboardingSummary }
-  | { type: "error"; data: string }
-  | { type: "complete"; data: Record<string, never> };
-
-/** Start a new planning session with an initial plan */
-export function startPlanning(
-  initialPlan: string,
-  projectId?: string,
-
-): Promise<PlanningSession> {
-  return api<PlanningSession>(withProjectId("/planning/start", projectId), {
-    method: "POST",
-    body: JSON.stringify({
-      initialPlan,
-    }),
-  });
-}
-
-export function createPlanningDraft(
-  initialPlan: string,
-  projectId?: string,
-  modelOverride?: { planningModelProvider?: string; planningModelId?: string; thinkingLevel?: ThinkingLevel },
-): Promise<{ sessionId: string; title: string }> {
-  return api<{ sessionId: string; title: string }>(withProjectId("/planning/create-draft", projectId), {
-    method: "POST",
-    body: JSON.stringify({
-      initialPlan,
-      planningModelProvider: modelOverride?.planningModelProvider,
-      planningModelId: modelOverride?.planningModelId,
-      thinkingLevel: modelOverride?.thinkingLevel,
-    }),
-  });
-}
-
-/** Start a new planning session with AI streaming support */
-export function startPlanningStreaming(
-  initialPlan: string,
-  projectId?: string,
-  modelOverride?: { planningModelProvider?: string; planningModelId?: string; thinkingLevel?: ThinkingLevel },
-  planningOptions?: { clarificationEnabled?: boolean; workflowId?: string | null },
-  existingSessionId?: string,
-): Promise<{ sessionId: string }> {
-  return api<{ sessionId: string }>(withProjectId("/planning/start-streaming", projectId), {
-    method: "POST",
-    body: JSON.stringify({
-      initialPlan,
-      planningModelProvider: modelOverride?.planningModelProvider,
-      planningModelId: modelOverride?.planningModelId,
-      thinkingLevel: modelOverride?.thinkingLevel,
-      clarificationEnabled: planningOptions?.clarificationEnabled,
-      ...(planningOptions?.workflowId ? { workflowId: planningOptions.workflowId } : {}),
-      ...(existingSessionId ? { existingSessionId } : {}),
-    }),
-  });
-}
-
-/** Explicitly validate the current running planning summary before creating work. */
-export function validatePlanningSession(sessionId: string, projectId?: string): Promise<{ summary: PlanningSummary; validated: boolean }> {
-  return api<{ summary: PlanningSummary; validated: boolean }>(withProjectId(`/planning/${encodeURIComponent(sessionId)}/validate`, projectId), { method: "POST" });
-}
-
-/** Rename a planning session after the server verifies the session type. */
-export function updatePlanningSessionTitle(sessionId: string, title: string, projectId?: string): Promise<{ sessionId: string; title: string }> {
-  return api<{ sessionId: string; title: string }>(withProjectId(`/planning/${encodeURIComponent(sessionId)}/title`, projectId), {
-    method: "PATCH",
-    body: JSON.stringify({ title }),
-  });
-}
-
-/** Submit a response to the current planning question */
-export function respondToPlanning(
-  sessionId: string,
-  responses: Record<string, unknown>,
-  projectId?: string,
-): Promise<PlanningResponse> {
-  return api<PlanningResponse>(withProjectId("/planning/respond", projectId), {
-    method: "POST",
-    body: JSON.stringify({ sessionId, responses }),
-  });
-}
-
-/** Rewind a planning session to the previous answered question */
-export function rewindPlanningSession(
-  sessionId: string,
-  projectId?: string,
-  questionId?: string,
-): Promise<{ currentQuestion: PlanningQuestion; summary?: PlanningSummary; history: Array<{ question: PlanningQuestion; response: unknown; thinkingOutput?: string }> }> {
-  return api<{ currentQuestion: PlanningQuestion; summary?: PlanningSummary; history: Array<{ question: PlanningQuestion; response: unknown; thinkingOutput?: string }> }>(
-    withProjectId(`/planning/${encodeURIComponent(sessionId)}/back`, projectId),
-    {
-      method: "POST",
-      ...(questionId ? { body: JSON.stringify({ questionId }) } : {}),
-    },
-  );
-}
-
-/** Retry a failed planning session turn */
-export function retryPlanningSession(
-  sessionId: string,
-  projectId?: string,
-): Promise<{ success: boolean; sessionId: string }> {
-  return api<{ success: boolean; sessionId: string }>(
-    withProjectId(`/planning/${encodeURIComponent(sessionId)}/retry`, projectId),
-    {
-      method: "POST",
-    },
-  );
-}
-
-/** Stop in-flight planning generation for a session */
-export function stopPlanningGeneration(
-  sessionId: string,
-  projectId?: string,
-): Promise<{ success: boolean }> {
-  return api<{ success: boolean }>(
-    withProjectId(`/planning/${encodeURIComponent(sessionId)}/stop`, projectId),
-    {
-      method: "POST",
-    },
-  );
-}
-
-/** Cancel an active planning session */
-export function cancelPlanning(sessionId: string, projectId?: string): Promise<void> {
-  return api<void>(withProjectId("/planning/cancel", projectId), {
-    method: "POST",
-    body: JSON.stringify({ sessionId }),
-  });
-}
-
-export function startAgentOnboardingStreaming(
-  intent: string,
-  context: {
-    existingAgents: Array<{ id: string; name: string; role: string }>;
-    templates: Array<{ id: string; label: string; description?: string }>;
-    mode?: OnboardingMode;
-    existingAgentConfig?: ExistingAgentOnboardingConfig;
-  },
-  projectId?: string,
-  modelOverride?: { planningModelProvider?: string; planningModelId?: string },
-): Promise<{ sessionId: string }> {
-  return api<{ sessionId: string }>(withProjectId("/agents/onboarding/start-streaming", projectId), {
-    method: "POST",
-    body: JSON.stringify({
-      intent,
-      context,
-      mode: context.mode,
-      existingAgentConfig: context.existingAgentConfig,
-      planningModelProvider: modelOverride?.planningModelProvider,
-      planningModelId: modelOverride?.planningModelId,
-    }),
-  });
-}
-
-export function respondToAgentOnboarding(
-  sessionId: string,
-  responses: Record<string, unknown>,
-  projectId?: string,
-): Promise<{ type: "question" | "complete"; data: PlanningQuestion | AgentOnboardingSummary }> {
-  return api(withProjectId("/agents/onboarding/respond", projectId), {
-    method: "POST",
-    body: JSON.stringify({ sessionId, responses }),
-  });
-}
-
-export function retryAgentOnboardingSession(sessionId: string, projectId?: string): Promise<{ success: boolean; sessionId: string }> {
-  return api(withProjectId(`/agents/onboarding/${encodeURIComponent(sessionId)}/retry`, projectId), {
-    method: "POST",
-  });
-}
-
-export function stopAgentOnboardingGeneration(sessionId: string, projectId?: string): Promise<{ success: boolean }> {
-  return api(withProjectId(`/agents/onboarding/${encodeURIComponent(sessionId)}/stop`, projectId), {
-    method: "POST",
-  });
-}
-
-export function cancelAgentOnboarding(sessionId: string, projectId?: string): Promise<void> {
-  return api(withProjectId("/agents/onboarding/cancel", projectId), {
-    method: "POST",
-    body: JSON.stringify({ sessionId }),
-  });
-}
-
-/** Create a task from a completed planning session */
-export function createTaskFromPlanning(
-  sessionId: string,
-  summary?: PlanningSummary,
-  projectId?: string,
-  options?: {
-    branch?: string;
-    baseBranch?: string;
-    branchSelection?: {
-      mode: "project-default" | "auto-new" | "existing" | "custom-new";
-      branchName?: string;
-      baseBranch?: string;
-    };
-    workflowId?: string | null;
-  },
-): Promise<Task> {
-  return api<{ task: Task; alreadyCreated: boolean }>(withProjectId("/planning/create-task", projectId), {
-    method: "POST",
-    body: JSON.stringify({
-      ...(summary ? { sessionId, summary } : { sessionId }),
-      ...(options?.branch !== undefined ? { branch: options.branch } : {}),
-      ...(options?.baseBranch !== undefined ? { baseBranch: options.baseBranch } : {}),
-      ...(options?.branchSelection ? { branchSelection: options.branchSelection } : {}),
-      ...(options?.workflowId !== undefined ? { workflowId: options.workflowId } : {}),
-    }),
-  }).then((response) => response.task);
-}
-
-/** Start subtask breakdown from a completed planning session */
-export function startPlanningBreakdown(
-  sessionId: string,
-  summary?: PlanningSummary,
-  projectId?: string,
-): Promise<{ sessionId: string; subtasks: SubtaskItem[] }> {
-  return api<{ sessionId: string; subtasks: SubtaskItem[] }>(
-    withProjectId("/planning/start-breakdown", projectId),
-    {
-      method: "POST",
-      body: JSON.stringify(summary ? { sessionId, summary } : { sessionId }),
-    },
-  );
-}
-
-/** Create multiple tasks from a completed planning session */
-export function createTasksFromPlanning(
-  planningSessionId: string,
-  subtasks: PlanningSubtaskDraft[],
-  projectId?: string,
-  options?: {
-    branchSelection?: {
-      mode: "project-default" | "auto-new" | "existing" | "custom-new";
-      branchName?: string;
-      baseBranch?: string;
-    };
-    branchAssignment?: {
-      mode: "shared" | "per-task-derived";
-    };
-    workflowId?: string | null;
-  },
-): Promise<{ tasks: Task[] }> {
-  return api<{ tasks: Task[] }>(withProjectId("/planning/create-tasks", projectId), {
-    method: "POST",
-    body: JSON.stringify({
-      planningSessionId,
-      subtasks,
-      ...(options?.branchSelection ? { branchSelection: options.branchSelection } : {}),
-      ...(options?.branchAssignment ? { branchAssignment: options.branchAssignment } : {}),
-      ...(options?.workflowId !== undefined ? { workflowId: options.workflowId } : {}),
-    }),
-  });
-}
-
+/*
+ * FNXC:CodeOrganization 2026-07-20-14:00:
+ * Preserve legacy `planning` imports while implementations live in planning.js.
+ */
+export {
+  startPlanning,
+  createPlanningDraft,
+  startPlanningStreaming,
+  validatePlanningSession,
+  updatePlanningSessionTitle,
+  respondToPlanning,
+  rewindPlanningSession,
+  retryPlanningSession,
+  stopPlanningGeneration,
+  cancelPlanning,
+  startAgentOnboardingStreaming,
+  respondToAgentOnboarding,
+  retryAgentOnboardingSession,
+  stopAgentOnboardingGeneration,
+  cancelAgentOnboarding,
+  createTaskFromPlanning,
+  startPlanningBreakdown,
+  createTasksFromPlanning,
+  getPlanningStreamUrl,
+  getAgentOnboardingStreamUrl,
+  connectAgentOnboardingStream,
+  connectPlanningStream,
+} from "./planning.js";
+export type {
+  PlanningSession,
+  PlanningResponse,
+  PlanningStreamEvent,
+  AgentOnboardingSummary,
+  OnboardingMode,
+  ThinkingLevel,
+  ExistingAgentOnboardingConfig,
+  AgentOnboardingStreamEvent,
+} from "./planning.js";
 
 // FNXC:CodeOrganization 2026-07-19-12:00: SSE reconnect lives in event-source.ts.
 export type { StreamConnectionState, ResilientEventSourceOptions, ResilientEventHandlers } from "./event-source.js";
-import { createResilientEventSource } from "./event-source.js";
-import type { StreamConnectionState } from "./event-source.js";
-import { startKeepAlive } from "./ai-sessions.js";
 export { createResilientEventSource } from "./event-source.js";
 
-export interface DevServerCandidate {
-  scriptName: string;
-  command: string;
-  packagePath: string;
-  confidence: number;
-  name: string;
-  cwd: string;
-  source: string;
-  workspaceName?: string;
-  label: string;
-}
-
-// Backward-compatible alias for backend naming in FN-2178 scope.
-export type DetectedCandidate = DevServerCandidate;
-
-export interface DevServerState {
-  id: string;
-  name: string;
-  status: "stopped" | "starting" | "running" | "failed";
-  command: string;
-  scriptName: string;
-  cwd: string;
-  pid?: number;
-  startedAt?: string;
-  previewUrl?: string;
-  detectedUrl?: string;
-  detectedPort?: number;
-  manualPreviewUrl?: string;
-  manualUrl?: string;
-  logs: string[];
-  exitCode?: number | null;
-}
-
-export type DevServerStatus = DevServerState;
-
-export interface DevServerStartInput {
-  command: string;
-  scriptName?: string;
-  cwd?: string;
-  packagePath?: string;
-}
-
-export interface DevServerConfig {
-  selectedScript: string | null;
-  selectedSource: string | null;
-  selectedCommand: string | null;
-  previewUrlOverride: string | null;
-  detectedPreviewUrl: string | null;
-  selectedAt: string | null;
-}
-
-export interface DevServerLogHistoryEntry {
-  id: number;
-  text: string;
-  stream: "stdout" | "stderr";
-  timestamp: string;
-}
-
-export interface DevServerLogHistoryResponse {
-  lines: DevServerLogHistoryEntry[];
-  totalLines: number;
-}
-
-export interface FetchDevServerLogHistoryOptions {
-  maxLines?: number;
-  offset?: number;
-  lastEventId?: number;
-}
-
-export interface DevServerConfig {
-  selectedScript: string | null;
-  selectedSource: string | null;
-  selectedCommand: string | null;
-  previewUrlOverride: string | null;
-  detectedPreviewUrl: string | null;
-  selectedAt: string | null;
-}
-
-interface BackendDevServerCandidate {
-  name: string;
-  command: string;
-  source?: string;
-  packageName?: string;
-  packagePath?: string;
-  confidence?: number;
-}
-
-interface BackendDevServerState {
-  id?: string;
-  name?: string;
-  status?: "stopped" | "starting" | "running" | "failed";
-  command?: string;
-  scriptId?: string;
-  cwd?: string;
-  pid?: number;
-  startedAt?: string;
-  previewUrl?: string;
-  detectedUrl?: string;
-  detectedPort?: number;
-  manualPreviewUrl?: string;
-  manualUrl?: string;
-  logHistory?: string[];
-  exitCode?: number | null;
-}
-
-interface BackendDevServerLogHistoryLine {
-  id?: number;
-  text?: string;
-  line?: string;
-  stream?: "stdout" | "stderr";
-  timestamp?: string;
-}
-
-interface BackendDevServerLogHistoryResponse {
-  lines?: BackendDevServerLogHistoryLine[];
-  totalLines?: number;
-}
-
-function mapBackendCandidateToFrontend(candidate: BackendDevServerCandidate): DevServerCandidate {
-  const source = typeof candidate.source === "string" && candidate.source.trim().length > 0
-    ? candidate.source.trim()
-    : "root";
-  const cwd = source === "root" ? "." : source;
-  const scriptName = candidate.name;
-  const packagePath = typeof candidate.packagePath === "string" && candidate.packagePath.trim().length > 0
-    ? candidate.packagePath.trim()
-    : cwd;
-  const confidence = typeof candidate.confidence === "number"
-    ? candidate.confidence
-    : 1;
-
-  const locationLabel = source === "root" ? "root" : source;
-  const packageLabel = typeof candidate.packageName === "string" && candidate.packageName.trim().length > 0
-    ? candidate.packageName.trim()
-    : "project";
-
-  return {
-    name: candidate.name,
-    command: candidate.command,
-    scriptName,
-    packagePath,
-    confidence,
-    cwd,
-    source,
-    workspaceName: typeof candidate.packageName === "string" ? candidate.packageName : undefined,
-    label: `${packageLabel} · ${scriptName} (${locationLabel})`,
-  };
-}
-
-function mapBackendStateToFrontend(state: BackendDevServerState): DevServerState {
-  const status = state.status;
-  const normalizedStatus = status === "starting" || status === "running" || status === "failed" || status === "stopped"
-    ? status
-    : "stopped";
-
-  const previewUrl = typeof state.previewUrl === "string"
-    ? state.previewUrl
-    : state.detectedUrl;
-  const manualPreviewUrl = typeof state.manualPreviewUrl === "string"
-    ? state.manualPreviewUrl
-    : state.manualUrl;
-
-  return {
-    id: typeof state.id === "string" ? state.id : "",
-    name: typeof state.name === "string" && state.name.length > 0 ? state.name : "default",
-    status: normalizedStatus,
-    command: typeof state.command === "string" ? state.command : "",
-    scriptName: typeof state.scriptId === "string" ? state.scriptId : "",
-    cwd: typeof state.cwd === "string" ? state.cwd : "",
-    pid: state.pid,
-    startedAt: state.startedAt,
-    previewUrl,
-    detectedUrl: typeof state.detectedUrl === "string" ? state.detectedUrl : previewUrl,
-    detectedPort: state.detectedPort,
-    manualPreviewUrl,
-    manualUrl: typeof state.manualUrl === "string" ? state.manualUrl : manualPreviewUrl,
-    logs: Array.isArray(state.logHistory) ? state.logHistory : [],
-    exitCode: state.exitCode,
-  };
-}
-
-function normalizeDevServerLogLine(line: BackendDevServerLogHistoryLine, fallbackId: number): DevServerLogHistoryEntry {
-  return {
-    id: typeof line.id === "number" && Number.isFinite(line.id) ? line.id : fallbackId,
-    text: typeof line.text === "string" ? line.text : (typeof line.line === "string" ? line.line : ""),
-    stream: line.stream === "stderr" ? "stderr" : "stdout",
-    timestamp: typeof line.timestamp === "string" ? line.timestamp : "",
-  };
-}
-
-function normalizeDevServerLogHistoryResponse(response: BackendDevServerLogHistoryResponse): DevServerLogHistoryResponse {
-  const rawLines = Array.isArray(response.lines) ? response.lines : [];
-  const lines = rawLines.map((line, index) => normalizeDevServerLogLine(line, index + 1));
-
-  return {
-    lines,
-    totalLines: typeof response.totalLines === "number" && Number.isFinite(response.totalLines)
-      ? response.totalLines
-      : lines.length,
-  };
-}
-
-function mapLegacyDevServerLogs(logs: string[], options: FetchDevServerLogHistoryOptions): DevServerLogHistoryResponse {
-  const maxLines = typeof options.maxLines === "number" && Number.isFinite(options.maxLines)
-    ? Math.max(1, Math.floor(options.maxLines))
-    : 100;
-  const offset = typeof options.offset === "number" && Number.isFinite(options.offset)
-    ? Math.max(0, Math.floor(options.offset))
-    : 0;
-  const lastEventId = typeof options.lastEventId === "number" && Number.isFinite(options.lastEventId)
-    ? Math.max(0, Math.floor(options.lastEventId))
-    : null;
-
-  const totalLines = logs.length;
-  const fullLines = logs.map<DevServerLogHistoryEntry>((text, index) => ({
-    id: index + 1,
-    text,
-    stream: "stdout",
-    timestamp: "",
-  }));
-
-  if (lastEventId !== null) {
-    return {
-      lines: fullLines.filter((line) => line.id > lastEventId).slice(0, maxLines),
-      totalLines,
-    };
-  }
-
-  const endExclusive = Math.max(totalLines - offset, 0);
-  const start = Math.max(endExclusive - maxLines, 0);
-
-  return {
-    lines: fullLines.slice(start, endExclusive),
-    totalLines,
-  };
-}
-
-type DevServerCandidatesResponse =
-  | { candidates?: BackendDevServerCandidate[] }
-  | BackendDevServerCandidate[];
-
-function mapCandidatesResponse(response: DevServerCandidatesResponse): DevServerCandidate[] {
-  if (Array.isArray(response)) {
-    return response.map(mapBackendCandidateToFrontend);
-  }
-
-  return (response.candidates ?? []).map(mapBackendCandidateToFrontend);
-}
-
-export async function fetchDevServerCandidates(projectId?: string): Promise<DevServerCandidate[]> {
-  try {
-    const response = await api<DevServerCandidatesResponse>(withProjectId("/dev-server/candidates", projectId));
-    return mapCandidatesResponse(response);
-  } catch (error) {
-    // Backward compatibility for workspaces that still expose /dev-server/detect.
-    if (error instanceof Error && /\/dev-server\/candidates/.test(error.message)) {
-      const fallback = await api<DevServerCandidatesResponse>(withProjectId("/dev-server/detect", projectId));
-      return mapCandidatesResponse(fallback);
-    }
-    throw error;
-  }
-}
-
-export function detectDevServer(projectId?: string): Promise<DevServerCandidate[]> {
-  return fetchDevServerCandidates(projectId);
-}
-
-export function fetchDevServerConfig(projectId?: string): Promise<DevServerConfig> {
-  return api<DevServerConfig>(withProjectId("/dev-server/config", projectId));
-}
-
-export function saveDevServerConfig(config: Partial<DevServerConfig>, projectId?: string): Promise<DevServerConfig> {
-  return api<DevServerConfig>(withProjectId("/dev-server/config", projectId), {
-    method: "PUT",
-    body: JSON.stringify(config),
-  });
-}
-
-export function fetchDevServerStatus(projectId?: string): Promise<DevServerState> {
-  return api<BackendDevServerState>(withProjectId("/dev-server/status", projectId)).then(mapBackendStateToFrontend);
-}
-
-export async function fetchDevServerLogHistory(
-  options: FetchDevServerLogHistoryOptions = {},
-  projectId?: string,
-): Promise<DevServerLogHistoryResponse> {
-  const query = new URLSearchParams();
-  if (typeof options.maxLines === "number" && Number.isFinite(options.maxLines)) {
-    query.set("maxLines", String(Math.max(1, Math.floor(options.maxLines))));
-  }
-  if (typeof options.offset === "number" && Number.isFinite(options.offset)) {
-    query.set("offset", String(Math.max(0, Math.floor(options.offset))));
-  }
-  if (typeof options.lastEventId === "number" && Number.isFinite(options.lastEventId)) {
-    query.set("lastEventId", String(Math.max(0, Math.floor(options.lastEventId))));
-  }
-
-  const suffix = query.size > 0 ? `?${query.toString()}` : "";
-
-  try {
-    const response = await api<BackendDevServerLogHistoryResponse>(
-      withProjectId(`/dev-server/logs/history${suffix}`, projectId),
-    );
-    return normalizeDevServerLogHistoryResponse(response);
-  } catch (error) {
-    // Backward compatibility for workspaces without /dev-server/logs/history.
-    if (error instanceof Error && /\/dev-server\/logs\/history/.test(error.message)) {
-      const status = await fetchDevServerStatus(projectId);
-      return mapLegacyDevServerLogs(status.logs, options);
-    }
-    throw error;
-  }
-}
-
-export function startDevServer(body: DevServerStartInput, projectId?: string): Promise<DevServerState> {
-  const cwd = body.cwd ?? body.packagePath ?? ".";
-  const scriptName = body.scriptName;
-
-  return api<BackendDevServerState>(withProjectId("/dev-server/start", projectId), {
-    method: "POST",
-    body: JSON.stringify({
-      command: body.command,
-      scriptName,
-      scriptId: scriptName,
-      cwd,
-      packagePath: body.packagePath,
-    }),
-  }).then(mapBackendStateToFrontend);
-}
-
-export function stopDevServer(projectId?: string): Promise<DevServerState> {
-  return api<BackendDevServerState>(withProjectId("/dev-server/stop", projectId), {
-    method: "POST",
-  }).then(mapBackendStateToFrontend);
-}
-
-export function restartDevServer(projectId?: string): Promise<DevServerState> {
-  return api<BackendDevServerState>(withProjectId("/dev-server/restart", projectId), {
-    method: "POST",
-  }).then(mapBackendStateToFrontend);
-}
-
-export async function setDevServerPreviewUrl(urlOrBody: string | { url: string | null }, projectId?: string): Promise<DevServerState> {
-  const body = typeof urlOrBody === "string"
-    ? { url: urlOrBody }
-    : urlOrBody;
-
-  try {
-    const response = await api<BackendDevServerState>(withProjectId("/dev-server/preview-url", projectId), {
-      method: "POST",
-      body: JSON.stringify(body),
-    });
-    return mapBackendStateToFrontend(response);
-  } catch (error) {
-    // Backward compatibility for workspaces that still use PUT.
-    if (error instanceof Error && /\/dev-server\/preview-url/.test(error.message)) {
-      const fallback = await api<BackendDevServerState>(withProjectId("/dev-server/preview-url", projectId), {
-        method: "PUT",
-        body: JSON.stringify(body),
-      });
-      return mapBackendStateToFrontend(fallback);
-    }
-    throw error;
-  }
-}
-
-export function getDevServerLogsStreamUrl(projectId?: string): string {
-  return buildApiUrl(withProjectId("/dev-server/logs/stream", projectId));
-}
-
-// =============================================================================
-// Session-based DevServer API (FN-2184 / FN-2185)
-// Target /api/devserver/* with fallback to /api/dev-server/* for migration safety
-// =============================================================================
-
-/**
- * Canonical session-based DevServer types.
- * These align with the new session model introduced in FN-2184.
+/*
+ * FNXC:CodeOrganization 2026-07-20-14:00:
+ * Preserve legacy `dev-server` imports while implementations live in dev-server.js.
  */
-
-// Detected dev server command (result of detectDevServerCommands)
-export interface DetectedDevServerCommand {
-  name: string;
-  command: string;
-  cwd: string;
-  scriptName: string;
-  packagePath: string;
-  framework?: string;
-}
-
-// Dev server log entry format
-export interface DevServerLogEntry {
-  timestamp: string;
-  stream: "stdout" | "stderr";
-  text: string;
-}
-
-// Preview URL response from backend
-export interface DevServerPreviewResponse {
-  url: string | null;
-  source: "auto" | "manual" | null;
-}
-
-// Dev server runtime info (process details)
-export interface DevServerRuntime {
-  pid: number;
-  startedAt: string;
-  exitCode?: number;
-  previewUrl?: string;
-}
-
-// Dev server configuration (saved settings)
-export interface DevServerSessionConfig {
-  id: string;
-  name: string;
-  command: string;
-  cwd: string;
-  env?: Record<string, string>;
-  autoStart?: boolean;
-}
-
-// Full DevServer session combining config, status, runtime, and logs
-export interface DevServerSession {
-  config: DevServerSessionConfig;
-  status: "stopped" | "starting" | "running" | "failed" | "stopping";
-  runtime?: DevServerRuntime;
-  previewUrl?: string;
-  logHistory: DevServerLogEntry[];
-}
-
-// Options for fetching log history
-export interface FetchDevServerLogsOptions {
-  maxLines?: number;
-  offset?: number;
-  lastEventId?: number;
-}
-
-// Backend response shape for log history
-interface BackendSessionLogResponse {
-  lines?: DevServerLogEntry[];
-  totalLines?: number;
-}
-
-// Backend response for preview endpoint
-interface BackendPreviewResponse {
-  url?: string | null;
-  source?: string | null;
-}
-
-// Backend response for list sessions
-interface BackendSessionsListResponse {
-  sessions?: DevServerSession[];
-}
-
-// Backend response for detect commands
-interface BackendDetectCommandsResponse {
-  candidates?: DetectedDevServerCommand[];
-}
-
-/**
- * Fetch all dev server sessions.
- * Targets /api/devserver with fallback to /api/dev-server (legacy compatibility).
- */
-export async function fetchDevServers(projectId?: string): Promise<DevServerSession[]> {
-  try {
-    const response = await api<BackendSessionsListResponse>(withProjectId("/devserver", projectId));
-    return response.sessions ?? [];
-  } catch {
-    // Fallback: try to get the legacy single-server state and wrap it in session format
-    try {
-      const legacy = await fetchDevServerStatus(projectId);
-      // Convert legacy state to session format
-      const session: DevServerSession = {
-        config: {
-          id: legacy.id ?? "default",
-          name: legacy.name ?? "Dev Server",
-          command: legacy.command ?? "",
-          cwd: legacy.cwd ?? ".",
-        },
-        status: legacy.status,
-        runtime: legacy.pid
-          ? {
-            pid: legacy.pid,
-            startedAt: legacy.startedAt ?? new Date().toISOString(),
-            exitCode: legacy.exitCode ?? undefined,
-            previewUrl: legacy.previewUrl,
-          }
-          : undefined,
-        previewUrl: legacy.previewUrl ?? legacy.detectedUrl ?? undefined,
-        logHistory: (legacy.logs ?? []).map<DevServerLogEntry>((text) => ({
-          timestamp: new Date().toISOString(),
-          stream: text.startsWith("[stderr]") ? "stderr" : "stdout",
-          text: text.replace(/^\[stderr\]\s*/, ""),
-        })),
-      };
-      return [session];
-    } catch {
-      return [];
-    }
-  }
-}
-
-/**
- * Create a new dev server session.
- * Targets /api/devserver with fallback to /api/dev-server/start (legacy compatibility).
- */
-export async function createDevServer(
-  data: { command: string; cwd?: string; name?: string; env?: Record<string, string> },
-  projectId?: string,
-): Promise<DevServerSession> {
-  const body = {
-    command: data.command,
-    cwd: data.cwd ?? ".",
-    name: data.name,
-    env: data.env,
-  };
-
-  try {
-    return await api<DevServerSession>(withProjectId("/devserver", projectId), {
-      method: "POST",
-      body: JSON.stringify(body),
-    });
-  } catch {
-    // Fallback: use legacy start endpoint
-    const legacy = await startDevServer({ command: data.command, cwd: data.cwd }, projectId);
-    return {
-      config: {
-        id: legacy.id ?? "default",
-        name: legacy.name ?? data.name ?? "Dev Server",
-        command: legacy.command,
-        cwd: legacy.cwd ?? data.cwd ?? ".",
-      },
-      status: legacy.status,
-      runtime: legacy.pid
-        ? {
-          pid: legacy.pid,
-          startedAt: legacy.startedAt ?? new Date().toISOString(),
-          exitCode: legacy.exitCode ?? undefined,
-          previewUrl: legacy.previewUrl,
-        }
-        : undefined,
-      previewUrl: legacy.previewUrl ?? legacy.detectedUrl ?? undefined,
-      logHistory: (legacy.logs ?? []).map<DevServerLogEntry>((text) => ({
-        timestamp: new Date().toISOString(),
-        stream: text.startsWith("[stderr]") ? "stderr" : "stdout",
-        text: text.replace(/^\[stderr\]\s*/, ""),
-      })),
-    };
-  }
-}
-
-/**
- * Fetch a specific dev server session by ID.
- * Targets /api/devserver/:id with fallback to /api/dev-server/status (legacy compatibility).
- */
-export async function fetchDevServer(id: string, projectId?: string): Promise<DevServerSession | null> {
-  try {
-    return await api<DevServerSession>(withProjectId(`/devserver/${encodeURIComponent(id)}`, projectId));
-  } catch {
-    // Fallback: try legacy status endpoint (single-server model)
-    try {
-      const legacy = await fetchDevServerStatus(projectId);
-      // If no ID or ID matches default, return legacy state as session
-      if (!id || id === "default" || id === legacy.id) {
-        return {
-          config: {
-            id: legacy.id ?? "default",
-            name: legacy.name ?? "Dev Server",
-            command: legacy.command ?? "",
-            cwd: legacy.cwd ?? ".",
-          },
-          status: legacy.status,
-          runtime: legacy.pid
-            ? {
-              pid: legacy.pid,
-              startedAt: legacy.startedAt ?? new Date().toISOString(),
-              exitCode: legacy.exitCode ?? undefined,
-              previewUrl: legacy.previewUrl,
-            }
-            : undefined,
-          previewUrl: legacy.previewUrl ?? legacy.detectedUrl ?? undefined,
-          logHistory: (legacy.logs ?? []).map<DevServerLogEntry>((text) => ({
-            timestamp: new Date().toISOString(),
-            stream: text.startsWith("[stderr]") ? "stderr" : "stdout",
-            text: text.replace(/^\[stderr\]\s*/, ""),
-          })),
-        };
-      }
-      return null;
-    } catch {
-      return null;
-    }
-  }
-}
-
-/**
- * Start a specific dev server by ID.
- * Targets /api/devserver/:id/start with fallback to /api/dev-server/start (legacy compatibility).
- */
-export async function startDevServerById(id: string, projectId?: string): Promise<DevServerSession> {
-  try {
-    return await api<DevServerSession>(withProjectId(`/devserver/${encodeURIComponent(id)}/start`, projectId), {
-      method: "POST",
-    });
-  } catch {
-    // Fallback: use legacy start endpoint (single-server model)
-    const legacy = await startDevServer({ command: "" }, projectId);
-    return {
-      config: {
-        id: legacy.id ?? id,
-        name: legacy.name ?? "Dev Server",
-        command: legacy.command ?? "",
-        cwd: legacy.cwd ?? ".",
-      },
-      status: legacy.status,
-      runtime: legacy.pid
-        ? {
-          pid: legacy.pid,
-          startedAt: legacy.startedAt ?? new Date().toISOString(),
-          exitCode: legacy.exitCode ?? undefined,
-          previewUrl: legacy.previewUrl,
-        }
-        : undefined,
-      previewUrl: legacy.previewUrl ?? legacy.detectedUrl ?? undefined,
-      logHistory: (legacy.logs ?? []).map<DevServerLogEntry>((text) => ({
-        timestamp: new Date().toISOString(),
-        stream: text.startsWith("[stderr]") ? "stderr" : "stdout",
-        text: text.replace(/^\[stderr\]\s*/, ""),
-      })),
-    };
-  }
-}
-
-/**
- * Stop a specific dev server by ID.
- * Targets /api/devserver/:id/stop with fallback to /api/dev-server/stop (legacy compatibility).
- */
-export async function stopDevServerById(id: string, projectId?: string): Promise<DevServerSession> {
-  try {
-    return await api<DevServerSession>(withProjectId(`/devserver/${encodeURIComponent(id)}/stop`, projectId), {
-      method: "POST",
-    });
-  } catch {
-    // Fallback: use legacy stop endpoint
-    const legacy = await stopDevServer(projectId);
-    return {
-      config: {
-        id: legacy.id ?? id,
-        name: legacy.name ?? "Dev Server",
-        command: legacy.command ?? "",
-        cwd: legacy.cwd ?? ".",
-      },
-      status: legacy.status,
-      runtime: legacy.pid
-        ? {
-          pid: legacy.pid,
-          startedAt: legacy.startedAt ?? new Date().toISOString(),
-          exitCode: legacy.exitCode ?? undefined,
-          previewUrl: legacy.previewUrl,
-        }
-        : undefined,
-      previewUrl: legacy.previewUrl ?? legacy.detectedUrl ?? undefined,
-      logHistory: (legacy.logs ?? []).map<DevServerLogEntry>((text) => ({
-        timestamp: new Date().toISOString(),
-        stream: text.startsWith("[stderr]") ? "stderr" : "stdout",
-        text: text.replace(/^\[stderr\]\s*/, ""),
-      })),
-    };
-  }
-}
-
-/**
- * Restart a specific dev server by ID.
- * Targets /api/devserver/:id/restart with fallback to /api/dev-server/restart (legacy compatibility).
- */
-export async function restartDevServerById(id: string, projectId?: string): Promise<DevServerSession> {
-  try {
-    return await api<DevServerSession>(withProjectId(`/devserver/${encodeURIComponent(id)}/restart`, projectId), {
-      method: "POST",
-    });
-  } catch {
-    // Fallback: use legacy restart endpoint
-    const legacy = await restartDevServer(projectId);
-    return {
-      config: {
-        id: legacy.id ?? id,
-        name: legacy.name ?? "Dev Server",
-        command: legacy.command ?? "",
-        cwd: legacy.cwd ?? ".",
-      },
-      status: legacy.status,
-      runtime: legacy.pid
-        ? {
-          pid: legacy.pid,
-          startedAt: legacy.startedAt ?? new Date().toISOString(),
-          exitCode: legacy.exitCode ?? undefined,
-          previewUrl: legacy.previewUrl,
-        }
-        : undefined,
-      previewUrl: legacy.previewUrl ?? legacy.detectedUrl ?? undefined,
-      logHistory: (legacy.logs ?? []).map<DevServerLogEntry>((text) => ({
-        timestamp: new Date().toISOString(),
-        stream: text.startsWith("[stderr]") ? "stderr" : "stdout",
-        text: text.replace(/^\[stderr\]\s*/, ""),
-      })),
-    };
-  }
-}
-
-/**
- * Delete a specific dev server by ID.
- * Targets /api/devserver/:id with fallback (no legacy equivalent).
- */
-export async function deleteDevServer(id: string, projectId?: string): Promise<void> {
-  try {
-    await api<void>(withProjectId(`/devserver/${encodeURIComponent(id)}`, projectId), {
-      method: "DELETE",
-    });
-  } catch {
-    // No fallback for delete in legacy API (single-server model)
-    // Silently ignore - deletion may not be supported in legacy mode
-  }
-}
-
-/**
- * Fetch logs for a specific dev server by ID.
- * Targets /api/devserver/:id/logs with fallback to /api/dev-server/logs/history (legacy compatibility).
- */
-export async function fetchDevServerLogs(
-  id: string,
-  opts: FetchDevServerLogsOptions = {},
-  projectId?: string,
-): Promise<{ lines: DevServerLogEntry[]; totalLines: number }> {
-  const query = new URLSearchParams();
-  if (typeof opts.maxLines === "number" && Number.isFinite(opts.maxLines)) {
-    query.set("maxLines", String(Math.max(1, Math.floor(opts.maxLines))));
-  }
-  if (typeof opts.offset === "number" && Number.isFinite(opts.offset)) {
-    query.set("offset", String(Math.max(0, Math.floor(opts.offset))));
-  }
-  if (typeof opts.lastEventId === "number" && Number.isFinite(opts.lastEventId)) {
-    query.set("lastEventId", String(Math.max(0, Math.floor(opts.lastEventId))));
-  }
-  const suffix = query.size > 0 ? `?${query.toString()}` : "";
-
-  try {
-    const response = await api<BackendSessionLogResponse>(
-      withProjectId(`/devserver/${encodeURIComponent(id)}/logs${suffix}`, projectId),
-    );
-    return {
-      lines: response.lines ?? [],
-      totalLines: response.totalLines ?? response.lines?.length ?? 0,
-    };
-  } catch {
-    // Fallback: use legacy log history endpoint
-    try {
-      const response = await fetchDevServerLogHistory(opts, projectId);
-      return {
-        lines: response.lines.map<DevServerLogEntry>((entry) => ({
-          timestamp: entry.timestamp,
-          stream: entry.stream,
-          text: entry.text,
-        })),
-        totalLines: response.totalLines,
-      };
-    } catch {
-      return { lines: [], totalLines: 0 };
-    }
-  }
-}
-
-/**
- * Fetch preview URL for a specific dev server by ID.
- * Targets /api/devserver/:id/preview with fallback to /api/dev-server/status (legacy compatibility).
- */
-export async function fetchDevServerPreview(id: string, projectId?: string): Promise<DevServerPreviewResponse> {
-  try {
-    const response = await api<BackendPreviewResponse>(
-      withProjectId(`/devserver/${encodeURIComponent(id)}/preview`, projectId),
-    );
-    return {
-      url: response.url ?? null,
-      source: (response.source as DevServerPreviewResponse["source"]) ?? null,
-    };
-  } catch {
-    // Fallback: use legacy status endpoint
-    try {
-      const legacy = await fetchDevServerStatus(projectId);
-      return {
-        url: legacy.previewUrl ?? legacy.detectedUrl ?? legacy.manualUrl ?? null,
-        source: legacy.manualUrl ? "manual" : "auto",
-      };
-    } catch {
-      return { url: null, source: null };
-    }
-  }
-}
-
-/**
- * Set preview URL for a specific dev server by ID.
- * Targets /api/devserver/:id/preview with fallback to /api/dev-server/preview-url (legacy compatibility).
- */
-export async function setDevServerPreviewUrlById(
-  id: string,
-  url: string | null,
-  projectId?: string,
-): Promise<DevServerPreviewResponse> {
-  try {
-    const response = await api<BackendPreviewResponse>(
-      withProjectId(`/devserver/${encodeURIComponent(id)}/preview`, projectId),
-      {
-        method: "POST",
-        body: JSON.stringify({ url }),
-      },
-    );
-    return {
-      url: response.url ?? null,
-      source: (response.source as DevServerPreviewResponse["source"]) ?? null,
-    };
-  } catch {
-    // Fallback: use legacy preview URL endpoint
-    const legacy = await setDevServerPreviewUrl({ url }, projectId);
-    return {
-      url: legacy.previewUrl ?? legacy.manualUrl ?? null,
-      source: "manual",
-    };
-  }
-}
-
-/**
- * Detect available dev server commands.
- * Targets /api/devserver/detect with fallback to /api/dev-server/detect (legacy compatibility).
- */
-export async function detectDevServerCommands(projectId?: string): Promise<DetectedDevServerCommand[]> {
-  try {
-    const response = await api<BackendDetectCommandsResponse>(withProjectId("/devserver/detect", projectId));
-    return response.candidates ?? [];
-  } catch {
-    // Fallback: use legacy detect endpoint
-    try {
-      const legacy = await fetchDevServerCandidates(projectId);
-      return legacy.map<DetectedDevServerCommand>((candidate) => ({
-        name: candidate.name,
-        command: candidate.command,
-        cwd: candidate.cwd,
-        scriptName: candidate.scriptName,
-        packagePath: candidate.packagePath,
-      }));
-    } catch {
-      return [];
-    }
-  }
-}
-
-/**
- * Get the SSE stream URL for a specific dev server session's logs.
- * Targets /api/devserver/:id/logs/stream with fallback to /api/dev-server/logs/stream (legacy compatibility).
- */
-export function getDevServerSessionLogsStreamUrl(id: string, projectId?: string): string {
-  // Try new session-scoped endpoint first
-  return buildApiUrl(withProjectId(`/devserver/${encodeURIComponent(id)}/logs/stream`, projectId));
-}
-
-
-/** Get the SSE stream URL for a planning session */
-export function getPlanningStreamUrl(sessionId: string, projectId?: string): string {
-  return buildApiUrl(withProjectId(`/planning/${encodeURIComponent(sessionId)}/stream`, projectId));
-}
-
-export function getAgentOnboardingStreamUrl(sessionId: string, projectId?: string): string {
-  return buildApiUrl(withProjectId(`/agents/onboarding/${encodeURIComponent(sessionId)}/stream`, projectId));
-}
-
-export function connectAgentOnboardingStream(
-  sessionId: string,
-  projectId: string | undefined,
-  handlers: {
-    onThinking?: (data: string) => void;
-    onQuestion?: (data: PlanningQuestion) => void;
-    onSummary?: (data: AgentOnboardingSummary) => void;
-    onError?: (data: string) => void;
-    onComplete?: () => void;
-    onConnectionStateChange?: (state: StreamConnectionState) => void;
-  },
-  options?: { maxReconnectAttempts?: number },
-): { close: () => void; isConnected: () => boolean } {
-  const url = getAgentOnboardingStreamUrl(sessionId, projectId);
-  const resilient = createResilientEventSource(
-    url,
-    {
-      events: {
-        thinking: (event) => {
-          try { handlers.onThinking?.(JSON.parse(event.data)); } catch { handlers.onThinking?.(event.data); }
-        },
-        question: (event) => {
-          try { handlers.onQuestion?.(JSON.parse(event.data) as PlanningQuestion); } catch { /* ignore parse error */ }
-        },
-        summary: (event) => {
-          try { handlers.onSummary?.(JSON.parse(event.data) as AgentOnboardingSummary); } catch { /* ignore parse error */ }
-        },
-        error: (event) => {
-          try {
-            const parsed = JSON.parse(event.data);
-            handlers.onError?.(parsed.message || parsed);
-          } catch {
-            handlers.onError?.(event.data || "Stream error");
-          }
-        },
-        complete: () => {
-          handlers.onComplete?.();
-        },
-      },
-    },
-    {
-      maxReconnectAttempts: options?.maxReconnectAttempts,
-      onConnectionStateChange: handlers.onConnectionStateChange,
-      onFatalError: (message) => handlers.onError?.(message),
-    },
-  );
-
-  return {
-    close: resilient.close,
-    isConnected: resilient.isConnected,
-  };
-}
-
-/** Connect to planning session SSE stream and handle events
- * 
- * Returns an object with:
- * - close: function to close the connection
- */
-export function connectPlanningStream(
-  sessionId: string,
-  projectId: string | undefined,
-  handlers: {
-    onThinking?: (data: string) => void;
-    onQuestion?: (data: PlanningQuestion) => void;
-    onSummary?: (data: PlanningSummary) => void;
-    onError?: (data: string) => void;
-    onComplete?: () => void;
-    onConnectionStateChange?: (state: StreamConnectionState) => void;
-  },
-  options?: { maxReconnectAttempts?: number },
-): { close: () => void; isConnected: () => boolean } {
-  const url = getPlanningStreamUrl(sessionId, projectId);
-  let keepAlive: { stop: () => void } | null = null;
-  let connection: { close: () => void; isConnected: () => boolean } | null = null;
-
-  const stopKeepAlive = () => {
-    keepAlive?.stop();
-    keepAlive = null;
-  };
-
-  const resilient = createResilientEventSource(
-    url,
-    {
-      onOpen: () => {
-        stopKeepAlive();
-        keepAlive = startKeepAlive(sessionId, projectId);
-      },
-      onMessage: (event) => {
-        if (event.data.startsWith(":")) return;
-      },
-      events: {
-        thinking: (event) => {
-          try {
-            handlers.onThinking?.(JSON.parse(event.data));
-          } catch {
-            handlers.onThinking?.(event.data);
-          }
-        },
-        question: (event) => {
-          try {
-            handlers.onQuestion?.(JSON.parse(event.data) as PlanningQuestion);
-          } catch (err) {
-            console.error("[planning] Failed to parse question event:", err);
-          }
-        },
-        summary: (event) => {
-          try {
-            handlers.onSummary?.(JSON.parse(event.data) as PlanningSummary);
-          } catch (err) {
-            console.error("[planning] Failed to parse summary event:", err);
-          }
-        },
-        error: (event) => {
-          try {
-            const parsed = JSON.parse(event.data);
-            handlers.onError?.(parsed.message || parsed);
-          } catch {
-            handlers.onError?.(event.data || "Stream error");
-          }
-          connection?.close();
-        },
-        complete: () => {
-          handlers.onComplete?.();
-          connection?.close();
-        },
-      },
-    },
-    {
-      maxReconnectAttempts: options?.maxReconnectAttempts,
-      onConnectionStateChange: handlers.onConnectionStateChange,
-      onFatalError: (message) => {
-        stopKeepAlive();
-        handlers.onError?.(message);
-      },
-    },
-  );
-
-  connection = {
-    close: () => {
-      stopKeepAlive();
-      resilient.close();
-    },
-    isConnected: resilient.isConnected,
-  };
-
-  return connection;
-}
+export {
+  fetchDevServerCandidates,
+  detectDevServer,
+  fetchDevServerConfig,
+  saveDevServerConfig,
+  fetchDevServerStatus,
+  fetchDevServerLogHistory,
+  startDevServer,
+  stopDevServer,
+  restartDevServer,
+  setDevServerPreviewUrl,
+  getDevServerLogsStreamUrl,
+  fetchDevServers,
+  createDevServer,
+  fetchDevServer,
+  startDevServerById,
+  stopDevServerById,
+  restartDevServerById,
+  deleteDevServer,
+  fetchDevServerLogs,
+  fetchDevServerPreview,
+  setDevServerPreviewUrlById,
+  detectDevServerCommands,
+  getDevServerSessionLogsStreamUrl,
+} from "./dev-server.js";
+export type {
+  DevServerCandidate,
+  DetectedCandidate,
+  DevServerState,
+  DevServerStatus,
+  DevServerStartInput,
+  DevServerConfig,
+  DevServerLogHistoryEntry,
+  DevServerLogHistoryResponse,
+  FetchDevServerLogHistoryOptions,
+  DetectedDevServerCommand,
+  DevServerLogEntry,
+  DevServerPreviewResponse,
+  DevServerRuntime,
+  DevServerSessionConfig,
+  DevServerSession,
+  FetchDevServerLogsOptions,
+} from "./dev-server.js";
 
 /*
  * FNXC:CodeOrganization 2026-07-19-12:00:
@@ -2247,65 +906,16 @@ export type {
   SettingsImportResponse,
 } from "./agent-import-generation.js";
 
-// --- AI Summarization API ---
-
-/** Response from title summarization endpoint */
-export interface SummarizeTitleResponse {
-  title: string;
-}
-
-/** Summarize a task description into a concise title using AI.
- * @param description - The task description to summarize (must be >200 chars; model input is truncated)
- * @param provider - Optional AI model provider (e.g., "anthropic")
- * @param modelId - Optional AI model ID (e.g., "claude-sonnet-4-5")
- * @param projectId - Optional project ID for scoped settings resolution
- * @returns The generated title (guaranteed ≤60 characters)
- * @throws Error with descriptive message for 400/429/503 errors
+/*
+ * FNXC:CodeOrganization 2026-07-20-14:00:
+ * Preserve legacy `ai-summarize` imports while implementations live in ai-summarize.js.
  */
-export async function summarizeTitle(
-  description: string,
-  provider?: string,
-  modelId?: string,
-  projectId?: string
-): Promise<string> {
-  const url = projectId
-    ? `/api/ai/summarize-title?projectId=${encodeURIComponent(projectId)}`
-    : "/api/ai/summarize-title";
-  const res = await fetch(url, {
-    method: "POST",
-    headers: withTokenHeader({ "Content-Type": "application/json" }),
-    body: JSON.stringify({ description, provider, modelId }),
-  });
-
-  const contentType = res.headers.get("content-type") ?? "";
-  const bodyText = await res.text();
-  const isJson = contentType.includes("application/json");
-
-  if (!isJson) {
-    throw new Error(`API returned non-JSON response: ${bodyText.slice(0, 100)}`);
-  }
-
-  const data = JSON.parse(bodyText) as { title?: string; error?: string };
-
-  if (!res.ok) {
-    const errorMessage = data.error || "Request failed";
-    if (res.status === 400) {
-      throw new Error(`Invalid request: ${errorMessage}`);
-    } else if (res.status === 429) {
-      throw new Error(`Rate limit exceeded: ${errorMessage}`);
-    } else if (res.status === 503) {
-      throw new Error(`AI service temporarily unavailable: ${errorMessage}`);
-    } else {
-      throw new Error(errorMessage);
-    }
-  }
-
-  if (!data.title) {
-    throw new Error("API returned empty title");
-  }
-
-  return data.title;
-}
+export {
+  summarizeTitle,
+} from "./ai-summarize.js";
+export type {
+  SummarizeTitleResponse,
+} from "./ai-summarize.js";
 
 /*
  * FNXC:CodeOrganization 2026-07-19-12:00:
@@ -2516,584 +1126,40 @@ export type {
   SliceWithFeatures,
   ValidationRunsResponse,
 } from "./missions.js";
-// FNXC:CodeOrganization 2026-07-18-16:30: re-export does not bind mission types locally; interview helpers still type against them.
-import type { Milestone, MissionWithHierarchy, Slice } from "./missions.js";
-
-// ── Mission Interview API ─────────────────────────────────────────────────
-
-/** Mission plan types returned by the interview AI */
-export interface MissionPlanFeature {
-  title: string;
-  description?: string;
-  acceptanceCriteria?: string;
-}
-
-export interface MissionPlanSlice {
-  title: string;
-  description?: string;
-  verification?: string;
-  features: MissionPlanFeature[];
-}
-
-export interface MissionPlanMilestone {
-  title: string;
-  description?: string;
-  verification?: string;
-  slices: MissionPlanSlice[];
-}
-
-export interface MissionPlanSummary {
-  missionTitle?: string;
-  missionDescription?: string;
-  milestones: MissionPlanMilestone[];
-}
-
-export type MissionInterviewResponse =
-  | { type: "question"; data: PlanningQuestion }
-  | { type: "complete"; data: MissionPlanSummary };
-
-/** Start a mission interview session with AI streaming */
-export function startMissionInterview(
-  missionTitle: string,
-  projectId?: string,
-  modelOverride?: { modelProvider?: string; modelId?: string; thinkingLevel?: ThinkingLevel },
-): Promise<{ sessionId: string }> {
-  return api<{ sessionId: string }>(withProjectId("/missions/interview/start", projectId), {
-    method: "POST",
-    body: JSON.stringify({
-      missionTitle,
-      modelProvider: modelOverride?.modelProvider,
-      modelId: modelOverride?.modelId,
-      thinkingLevel: modelOverride?.thinkingLevel,
-    }),
-  });
-}
-
-/** Submit a response to the current interview question */
-export function respondToMissionInterview(
-  sessionId: string,
-  responses: Record<string, unknown>,
-  projectId?: string,
-): Promise<MissionInterviewResponse> {
-  return api<MissionInterviewResponse>(withProjectId("/missions/interview/respond", projectId), {
-    method: "POST",
-    body: JSON.stringify({ sessionId, responses }),
-  });
-}
-
-/** Retry a failed mission interview turn */
-export function retryMissionInterviewSession(
-  sessionId: string,
-  projectId?: string,
-): Promise<{ success: boolean; sessionId: string }> {
-  return api<{ success: boolean; sessionId: string }>(
-    withProjectId(`/missions/interview/${encodeURIComponent(sessionId)}/retry`, projectId),
-    { method: "POST" },
-  );
-}
-
-/** Cancel an active mission interview session */
-export function cancelMissionInterview(sessionId: string, projectId?: string): Promise<void> {
-  return api<void>(withProjectId("/missions/interview/cancel", projectId), {
-    method: "POST",
-    body: JSON.stringify({ sessionId }),
-  });
-}
-
-export async function fetchMissionInterviewDrafts(projectId?: string): Promise<MissionInterviewDraftSummary[]> {
-  const query = projectId ? `?${new URLSearchParams({ projectId }).toString()}` : "";
-  const result = await api<{ drafts?: MissionInterviewDraftSummary[] }>(`/missions/interview/drafts${query}`);
-  return result.drafts ?? [];
-}
-
-export function discardMissionInterviewDraft(
-  sessionId: string,
-  projectId?: string,
-): Promise<{ removed: boolean }> {
-  return api<{ removed: boolean }>(
-    withProjectId(`/missions/interview/drafts/${encodeURIComponent(sessionId)}/discard`, projectId),
-    { method: "POST" },
-  );
-}
-
-/** Create mission from completed interview */
-export function createMissionFromInterview(
-  sessionId: string,
-  summary?: MissionPlanSummary,
-  projectId?: string,
-  options?: {
-    branch?: string;
-    baseBranch?: string;
-    branchSelection?: {
-      mode: "project-default" | "auto-new" | "existing" | "custom-new";
-      branchName?: string;
-      baseBranch?: string;
-    };
-    branchAssignment?: { mode: "shared" | "per-task-derived" };
-  },
-): Promise<MissionWithHierarchy> {
-  return api<MissionWithHierarchy>(withProjectId("/missions/interview/create-mission", projectId), {
-    method: "POST",
-    body: JSON.stringify({
-      sessionId,
-      summary,
-      ...(options?.branch !== undefined ? { branch: options.branch } : {}),
-      ...(options?.baseBranch !== undefined ? { baseBranch: options.baseBranch } : {}),
-      ...(options?.branchSelection ? { branchSelection: options.branchSelection } : {}),
-      ...(options?.branchAssignment ? { branchAssignment: options.branchAssignment } : {}),
-    }),
-  });
-}
-
-const MISSION_INTERVIEW_STREAM_ERROR_MESSAGE = "The mission interview stream was interrupted. Please retry the session.";
-
-function normalizeMissionInterviewStreamError(data: string | undefined): string {
-  const raw = data?.trim() ?? "";
-  if (!raw) return MISSION_INTERVIEW_STREAM_ERROR_MESSAGE;
-
-  const normalizeMessage = (value: unknown): string => {
-    if (typeof value !== "string") return MISSION_INTERVIEW_STREAM_ERROR_MESSAGE;
-    const message = value.trim();
-    if (!message || message === "Stream error") return MISSION_INTERVIEW_STREAM_ERROR_MESSAGE;
-    return message;
-  };
-
-  try {
-    const parsed = JSON.parse(raw) as unknown;
-    if (parsed && typeof parsed === "object" && !Array.isArray(parsed)) {
-      const message = (parsed as { message?: unknown; error?: unknown }).message ?? (parsed as { error?: unknown }).error;
-      return normalizeMessage(message);
-    }
-    return normalizeMessage(parsed);
-  } catch {
-    return normalizeMessage(raw);
-  }
-}
-
-/** Connect to mission interview SSE stream and handle events */
-export function connectMissionInterviewStream(
-  sessionId: string,
-  projectId: string | undefined,
-  handlers: {
-    onThinking?: (data: string) => void;
-    onQuestion?: (data: PlanningQuestion) => void;
-    onSummary?: (data: MissionPlanSummary) => void;
-    onError?: (data: string) => void;
-    onComplete?: () => void;
-    onConnectionStateChange?: (state: StreamConnectionState) => void;
-  },
-  options?: { maxReconnectAttempts?: number },
-): { close: () => void; isConnected: () => boolean } {
-  const url = buildApiUrl(withProjectId(`/missions/interview/${encodeURIComponent(sessionId)}/stream`, projectId));
-  let keepAlive: { stop: () => void } | null = null;
-  let connection: { close: () => void; isConnected: () => boolean } | null = null;
-  let terminalEventHandled = false;
-
-  const stopKeepAlive = () => {
-    keepAlive?.stop();
-    keepAlive = null;
-  };
-
-  const closeTerminalConnection = () => {
-    stopKeepAlive();
-    connection?.close();
-  };
-
-  const notifyTerminalError = (message: string) => {
-    if (terminalEventHandled) return;
-    terminalEventHandled = true;
-    closeTerminalConnection();
-    handlers.onError?.(message);
-  };
-
-  const notifyTerminalComplete = () => {
-    if (terminalEventHandled) return;
-    terminalEventHandled = true;
-    closeTerminalConnection();
-    handlers.onComplete?.();
-  };
-
-  const resilient = createResilientEventSource(
-    url,
-    {
-      onOpen: () => {
-        stopKeepAlive();
-        keepAlive = startKeepAlive(sessionId, projectId);
-      },
-      onMessage: (event) => {
-        if (event.data.startsWith(":")) return;
-      },
-      events: {
-        thinking: (event) => {
-          try {
-            handlers.onThinking?.(JSON.parse(event.data));
-          } catch {
-            handlers.onThinking?.(event.data);
-          }
-        },
-        question: (event) => {
-          try {
-            handlers.onQuestion?.(JSON.parse(event.data) as PlanningQuestion);
-          } catch (err) {
-            console.error("[mission-interview] Failed to parse question event:", err);
-          }
-        },
-        summary: (event) => {
-          try {
-            handlers.onSummary?.(JSON.parse(event.data) as MissionPlanSummary);
-          } catch (err) {
-            console.error("[mission-interview] Failed to parse summary event:", err);
-          }
-        },
-        error: (event) => {
-          /*
-          FNXC:MissionInterviewStream 2026-06-24-00:00:
-          Mission interview stream failures are terminal for the current EventSource. Normalize malformed/empty/generic payloads, close keepalive + SSE once, and ignore duplicate late error/complete events so the modal can show one recoverable Retry state instead of a stale spinner or raw stream failure.
-          */
-          notifyTerminalError(normalizeMissionInterviewStreamError(event.data));
-        },
-        complete: () => {
-          notifyTerminalComplete();
-        },
-      },
-    },
-    {
-      maxReconnectAttempts: options?.maxReconnectAttempts,
-      onConnectionStateChange: handlers.onConnectionStateChange,
-      onFatalError: (message) => {
-        notifyTerminalError(normalizeMissionInterviewStreamError(message));
-      },
-    },
-  );
-
-  connection = {
-    close: () => {
-      stopKeepAlive();
-      resilient.close();
-    },
-    isConnected: resilient.isConnected,
-  };
-
-  return connection;
-}
-
-// ── Milestone/Slice Interview API ─────────────────────────────────────────
-
-/** Summary type for milestone/slice interview responses */
-export interface TargetInterviewSummary {
-  title?: string;
-  description?: string;
-  planningNotes?: string;
-  verification?: string;
-}
-
-/** Response from milestone/slice interview: either a question or a completed plan */
-export type TargetInterviewResponse =
-  | { type: "question"; data: PlanningQuestion }
-  | { type: "complete"; data: TargetInterviewSummary };
-
-// Helper functions for URL construction
-function buildMilestoneInterviewUrl(milestoneId: string, path: string, projectId?: string): string {
-  return withProjectId(
-    `/missions/milestones/${encodeURIComponent(milestoneId)}/interview${path}`,
-    projectId
-  );
-}
-
-function buildSliceInterviewUrl(sliceId: string, path: string, projectId?: string): string {
-  return withProjectId(
-    `/missions/slices/${encodeURIComponent(sliceId)}/interview${path}`,
-    projectId
-  );
-}
-
-/** Start a milestone interview session */
-export function startMilestoneInterview(
-  milestoneId: string,
-  projectId?: string,
-): Promise<{ sessionId: string }> {
-  return api<{ sessionId: string }>(buildMilestoneInterviewUrl(milestoneId, "/start", projectId), {
-    method: "POST",
-  });
-}
-
-/** Submit a response to a milestone interview question */
-export function respondToMilestoneInterview(
-  sessionId: string,
-  responses: Record<string, unknown>,
-  projectId?: string,
-): Promise<TargetInterviewResponse> {
-  return api<TargetInterviewResponse>(buildMilestoneInterviewUrl(sessionId, "/respond", projectId), {
-    method: "POST",
-    body: JSON.stringify({ sessionId, responses }),
-  });
-}
-
-/** Connect to milestone interview SSE stream and handle events */
-export function connectMilestoneInterviewStream(
-  sessionId: string,
-  projectId: string | undefined,
-  handlers: {
-    onThinking?: (data: string) => void;
-    onQuestion?: (data: PlanningQuestion) => void;
-    onSummary?: (data: TargetInterviewSummary) => void;
-    onError?: (data: string) => void;
-    onComplete?: () => void;
-    onConnectionStateChange?: (state: StreamConnectionState) => void;
-  },
-  options?: { maxReconnectAttempts?: number },
-): { close: () => void; isConnected: () => boolean } {
-  const url = buildApiUrl(buildMilestoneInterviewUrl(sessionId, `/${encodeURIComponent(sessionId)}/stream`, projectId));
-  let keepAlive: { stop: () => void } | null = null;
-  let connection: { close: () => void; isConnected: () => boolean } | null = null;
-
-  const stopKeepAlive = () => {
-    keepAlive?.stop();
-    keepAlive = null;
-  };
-
-  const resilient = createResilientEventSource(
-    url,
-    {
-      onOpen: () => {
-        stopKeepAlive();
-        keepAlive = startKeepAlive(sessionId, projectId);
-      },
-      onMessage: (event) => {
-        if (event.data.startsWith(":")) return;
-      },
-      events: {
-        thinking: (event) => {
-          try {
-            handlers.onThinking?.(JSON.parse(event.data));
-          } catch {
-            handlers.onThinking?.(event.data);
-          }
-        },
-        question: (event) => {
-          try {
-            handlers.onQuestion?.(JSON.parse(event.data) as PlanningQuestion);
-          } catch (err) {
-            console.error("[milestone-interview] Failed to parse question event:", err);
-          }
-        },
-        summary: (event) => {
-          try {
-            handlers.onSummary?.(JSON.parse(event.data) as TargetInterviewSummary);
-          } catch (err) {
-            console.error("[milestone-interview] Failed to parse summary event:", err);
-          }
-        },
-        error: (event) => {
-          try {
-            const parsed = JSON.parse(event.data);
-            handlers.onError?.(parsed.message || parsed);
-          } catch {
-            handlers.onError?.(event.data || "Stream error");
-          }
-          connection?.close();
-        },
-        complete: () => {
-          handlers.onComplete?.();
-          connection?.close();
-        },
-      },
-    },
-    {
-      maxReconnectAttempts: options?.maxReconnectAttempts,
-      onConnectionStateChange: handlers.onConnectionStateChange,
-      onFatalError: (message) => {
-        stopKeepAlive();
-        handlers.onError?.(message);
-      },
-    },
-  );
-
-  connection = {
-    close: () => {
-      stopKeepAlive();
-      resilient.close();
-    },
-    isConnected: resilient.isConnected,
-  };
-
-  return connection;
-}
-
-/** Apply milestone interview results to the milestone */
-export function applyMilestoneInterview(
-  sessionId: string,
-  summary?: TargetInterviewSummary,
-  projectId?: string,
-): Promise<Milestone> {
-  return api<Milestone>(buildMilestoneInterviewUrl(sessionId, "/apply", projectId), {
-    method: "POST",
-    body: JSON.stringify({ sessionId, summary }),
-  });
-}
-
-/** Skip milestone interview and use mission context */
-export function skipMilestoneInterview(
-  milestoneId: string,
-  projectId?: string,
-): Promise<Milestone> {
-  return api<Milestone>(buildMilestoneInterviewUrl(milestoneId, "/skip", projectId), {
-    method: "POST",
-  });
-}
-
-/** Start a slice interview session */
-export function startSliceInterview(
-  sliceId: string,
-  projectId?: string,
-): Promise<{ sessionId: string }> {
-  return api<{ sessionId: string }>(buildSliceInterviewUrl(sliceId, "/start", projectId), {
-    method: "POST",
-  });
-}
-
-/** Submit a response to a slice interview question */
-export function respondToSliceInterview(
-  sessionId: string,
-  responses: Record<string, unknown>,
-  projectId?: string,
-): Promise<TargetInterviewResponse> {
-  return api<TargetInterviewResponse>(buildSliceInterviewUrl(sessionId, "/respond", projectId), {
-    method: "POST",
-    body: JSON.stringify({ sessionId, responses }),
-  });
-}
-
-/** Connect to slice interview SSE stream and handle events */
-export function connectSliceInterviewStream(
-  sessionId: string,
-  projectId: string | undefined,
-  handlers: {
-    onThinking?: (data: string) => void;
-    onQuestion?: (data: PlanningQuestion) => void;
-    onSummary?: (data: TargetInterviewSummary) => void;
-    onError?: (data: string) => void;
-    onComplete?: () => void;
-    onConnectionStateChange?: (state: StreamConnectionState) => void;
-  },
-  options?: { maxReconnectAttempts?: number },
-): { close: () => void; isConnected: () => boolean } {
-  const url = buildApiUrl(buildSliceInterviewUrl(sessionId, `/${encodeURIComponent(sessionId)}/stream`, projectId));
-  let keepAlive: { stop: () => void } | null = null;
-  let connection: { close: () => void; isConnected: () => boolean } | null = null;
-
-  const stopKeepAlive = () => {
-    keepAlive?.stop();
-    keepAlive = null;
-  };
-
-  const resilient = createResilientEventSource(
-    url,
-    {
-      onOpen: () => {
-        stopKeepAlive();
-        keepAlive = startKeepAlive(sessionId, projectId);
-      },
-      onMessage: (event) => {
-        if (event.data.startsWith(":")) return;
-      },
-      events: {
-        thinking: (event) => {
-          try {
-            handlers.onThinking?.(JSON.parse(event.data));
-          } catch {
-            handlers.onThinking?.(event.data);
-          }
-        },
-        question: (event) => {
-          try {
-            handlers.onQuestion?.(JSON.parse(event.data) as PlanningQuestion);
-          } catch (err) {
-            console.error("[slice-interview] Failed to parse question event:", err);
-          }
-        },
-        summary: (event) => {
-          try {
-            handlers.onSummary?.(JSON.parse(event.data) as TargetInterviewSummary);
-          } catch (err) {
-            console.error("[slice-interview] Failed to parse summary event:", err);
-          }
-        },
-        error: (event) => {
-          try {
-            const parsed = JSON.parse(event.data);
-            handlers.onError?.(parsed.message || parsed);
-          } catch {
-            handlers.onError?.(event.data || "Stream error");
-          }
-          connection?.close();
-        },
-        complete: () => {
-          handlers.onComplete?.();
-          connection?.close();
-        },
-      },
-    },
-    {
-      maxReconnectAttempts: options?.maxReconnectAttempts,
-      onConnectionStateChange: handlers.onConnectionStateChange,
-      onFatalError: (message) => {
-        stopKeepAlive();
-        handlers.onError?.(message);
-      },
-    },
-  );
-
-  connection = {
-    close: () => {
-      stopKeepAlive();
-      resilient.close();
-    },
-    isConnected: resilient.isConnected,
-  };
-
-  return connection;
-}
-
-/** Apply slice interview results to the slice */
-export function applySliceInterview(
-  sessionId: string,
-  summary?: TargetInterviewSummary,
-  projectId?: string,
-): Promise<Slice> {
-  return api<Slice>(buildSliceInterviewUrl(sessionId, "/apply", projectId), {
-    method: "POST",
-    body: JSON.stringify({ sessionId, summary }),
-  });
-}
-
-/** Skip slice interview and use mission context */
-export function skipSliceInterview(
-  sliceId: string,
-  projectId?: string,
-): Promise<Slice> {
-  return api<Slice>(buildSliceInterviewUrl(sliceId, "/skip", projectId), {
-    method: "POST",
-  });
-}
-
-/** Preview enriched description for a feature before triage */
-export async function previewEnrichedDescription(
-  featureId: string,
-  projectId?: string,
-): Promise<{ description: string }> {
-  try {
-    return await api<{ description: string }>(
-      withProjectId(`/missions/features/${encodeURIComponent(featureId)}/preview-description`, projectId),
-      {
-        method: "POST",
-      }
-    );
-  } catch {
-    // If endpoint doesn't exist, throw to trigger fallback
-    throw new Error("Preview endpoint not available");
-  }
-}
+/*
+ * FNXC:CodeOrganization 2026-07-20-14:00:
+ * Preserve legacy `mission-interview` imports while implementations live in mission-interview.js.
+ */
+export {
+  startMissionInterview,
+  respondToMissionInterview,
+  retryMissionInterviewSession,
+  cancelMissionInterview,
+  fetchMissionInterviewDrafts,
+  discardMissionInterviewDraft,
+  createMissionFromInterview,
+  connectMissionInterviewStream,
+  startMilestoneInterview,
+  respondToMilestoneInterview,
+  connectMilestoneInterviewStream,
+  applyMilestoneInterview,
+  skipMilestoneInterview,
+  startSliceInterview,
+  respondToSliceInterview,
+  connectSliceInterviewStream,
+  applySliceInterview,
+  skipSliceInterview,
+  previewEnrichedDescription,
+} from "./mission-interview.js";
+export type {
+  MissionPlanFeature,
+  MissionPlanSlice,
+  MissionPlanMilestone,
+  MissionPlanSummary,
+  MissionInterviewResponse,
+  TargetInterviewSummary,
+  TargetInterviewResponse,
+} from "./mission-interview.js";
 
 /*
  * FNXC:CodeOrganization 2026-07-18-14:00:
