+++ scripts/line-count-baseline.json
@@ -1,38 +1,37 @@
 {
   "packages/cli/src/__tests__/extension.test.ts": 4491,
-  "packages/cli/src/bin.ts": 2345,
+  "packages/cli/src/bin.ts": 2338,
   "packages/cli/src/commands/__tests__/serve.test.ts": 2313,
   "packages/cli/src/commands/__tests__/task.test.ts": 3644,
-  "packages/cli/src/commands/dashboard-tui/app.tsx": 4674,
-  "packages/cli/src/commands/dashboard.ts": 3704,
+  "packages/cli/src/commands/dashboard-tui/app.tsx": 4697,
+  "packages/cli/src/commands/dashboard.ts": 3729,
   "packages/cli/src/commands/task.ts": 2372,
   "packages/cli/src/extension.ts": 5694,
-  "packages/core/src/__tests__/postgres/schema-applier.test.ts": 2247,
+  "packages/core/src/__tests__/postgres/schema-applier.test.ts": 2435,
   "packages/core/src/agent-store.ts": 3484,
   "packages/core/src/async-mission-store-queries.ts": 2093,
   "packages/core/src/central-core.ts": 4486,
-  "packages/core/src/index.gate.ts": 2246,
-  "packages/core/src/index.ts": 2597,
+  "packages/core/src/index.gate.ts": 2248,
+  "packages/core/src/index.ts": 2610,
   "packages/core/src/mission-store.ts": 4369,
-  "packages/core/src/postgres/schema/project.ts": 2216,
-  "packages/core/src/postgres/sqlite-migrator.ts": 2298,
+  "packages/core/src/postgres/schema/project.ts": 2218,
+  "packages/core/src/postgres/sqlite-migrator.ts": 2326,
   "packages/core/src/store.ts": 2817,
-  "packages/core/src/types.ts": 6215,
+  "packages/core/src/types.ts": 6263,
   "packages/dashboard/app/App.tsx": 2110,
-  "packages/dashboard/app/api/legacy.ts": 3339,
   "packages/dashboard/app/components/AgentDetailView.tsx": 5567,
   "packages/dashboard/app/components/AgentsView.tsx": 2254,
-  "packages/dashboard/app/components/ChatView.tsx": 4024,
+  "packages/dashboard/app/components/ChatView.tsx": 4043,
   "packages/dashboard/app/components/GitHubImportModal.tsx": 2481,
   "packages/dashboard/app/components/GitManagerModal.tsx": 3560,
   "packages/dashboard/app/components/ListView.tsx": 3118,
   "packages/dashboard/app/components/MissionManager.tsx": 5149,
   "packages/dashboard/app/components/ModelOnboardingModal.tsx": 3594,
-  "packages/dashboard/app/components/PlanningModeModal.tsx": 3940,
+  "packages/dashboard/app/components/PlanningModeModal.tsx": 4538,
   "packages/dashboard/app/components/QuickEntryBox.tsx": 2562,
   "packages/dashboard/app/components/SettingsModal.tsx": 5258,
-  "packages/dashboard/app/components/TaskCard.tsx": 3818,
-  "packages/dashboard/app/components/TaskDetailModal.tsx": 6523,
+  "packages/dashboard/app/components/TaskCard.tsx": 3857,
+  "packages/dashboard/app/components/TaskDetailModal.tsx": 6528,
   "packages/dashboard/app/components/TerminalModal.tsx": 3322,
   "packages/dashboard/app/components/WorkflowNodeEditor.tsx": 5685,
   "packages/dashboard/app/components/__tests__/AgentsView.test.tsx": 2854,
@@ -42,17 +41,16 @@
   "packages/dashboard/app/components/__tests__/GitHubImportModal.test.tsx": 3171,
   "packages/dashboard/app/components/__tests__/GitManagerModal.test.tsx": 3790,
   "packages/dashboard/app/components/__tests__/ListView.test.tsx": 5380,
-  "packages/dashboard/app/components/__tests__/MailboxView.test.tsx": 2613,
+  "packages/dashboard/app/components/__tests__/MailboxView.test.tsx": 2646,
   "packages/dashboard/app/components/__tests__/ModelOnboardingModal.test.tsx": 5146,
   "packages/dashboard/app/components/__tests__/NewTaskModal.test.tsx": 2081,
-  "packages/dashboard/app/components/__tests__/PlanningModeModal.planning-flow.test.tsx": 3990,
   "packages/dashboard/app/components/__tests__/QuickEntryBox.test.tsx": 5433,
   "packages/dashboard/app/components/__tests__/SettingsModal.general.test.tsx": 2166,
-  "packages/dashboard/app/components/__tests__/TaskCard.test.tsx": 7551,
+  "packages/dashboard/app/components/__tests__/TaskCard.test.tsx": 7640,
   "packages/dashboard/app/components/__tests__/TaskChatTab.test.tsx": 3175,
   "packages/dashboard/app/components/__tests__/TaskDetailModal.inline-editing-and-integrations.test.tsx": 3438,
-  "packages/dashboard/app/components/__tests__/TaskDetailModal.rendering.test.tsx": 2846,
-  "packages/dashboard/app/components/__tests__/TaskDetailModal.responsive-and-dependencies.test.tsx": 2058,
+  "packages/dashboard/app/components/__tests__/TaskDetailModal.rendering.test.tsx": 2854,
+  "packages/dashboard/app/components/__tests__/TaskDetailModal.responsive-and-dependencies.test.tsx": 2071,
   "packages/dashboard/app/components/__tests__/TerminalModal.test.tsx": 9176,
   "packages/dashboard/app/components/__tests__/UsageIndicator.test.tsx": 2931,
   "packages/dashboard/app/components/__tests__/WorkflowNodeEditor.test.tsx": 4353,
@@ -60,20 +58,20 @@
   "packages/dashboard/app/components/__tests__/workflow-flow-mapping.test.ts": 2458,
   "packages/dashboard/app/hooks/__tests__/useChat.test.ts": 5466,
   "packages/dashboard/app/hooks/__tests__/useTasks.test.ts": 3766,
-  "packages/dashboard/src/__tests__/chat-manager.test.ts": 3807,
+  "packages/dashboard/src/__tests__/chat-manager.test.ts": 3859,
   "packages/dashboard/src/__tests__/file-service.test.ts": 2310,
   "packages/dashboard/src/__tests__/github.test.ts": 2467,
   "packages/dashboard/src/__tests__/routes-auth.test.ts": 5510,
   "packages/dashboard/src/__tests__/routes-automation.test.ts": 3089,
   "packages/dashboard/src/__tests__/routes-github.test.ts": 3900,
-  "packages/dashboard/src/__tests__/routes-planning.test.ts": 5321,
+  "packages/dashboard/src/__tests__/routes-planning.test.ts": 5379,
   "packages/dashboard/src/__tests__/routes-tasks.test.ts": 3024,
   "packages/dashboard/src/__tests__/usage.test.ts": 5261,
-  "packages/dashboard/src/chat.ts": 3130,
-  "packages/dashboard/src/github.ts": 5108,
+  "packages/dashboard/src/chat.ts": 3134,
+  "packages/dashboard/src/github.ts": 5112,
   "packages/dashboard/src/mission-routes.ts": 3924,
-  "packages/dashboard/src/planning.ts": 3341,
-  "packages/dashboard/src/routes.ts": 3272,
+  "packages/dashboard/src/planning.ts": 3640,
+  "packages/dashboard/src/routes.ts": 2043,
   "packages/dashboard/src/routes/register-git-github.ts": 6342,
   "packages/dashboard/src/routes/register-settings-memory-routes.ts": 2403,
   "packages/dashboard/src/routes/register-task-workflow-routes.ts": 5628,
@@ -89,20 +87,20 @@
   "packages/engine/src/__tests__/mission-execution-loop.test.ts": 2984,
   "packages/engine/src/__tests__/pi-create-fn-agent.test.ts": 3167,
   "packages/engine/src/__tests__/project-engine.test.ts": 3691,
-  "packages/engine/src/__tests__/restart.integration.test.ts": 2161,
-  "packages/engine/src/__tests__/self-healing.test.ts": 11779,
+  "packages/engine/src/__tests__/restart.integration.test.ts": 2199,
+  "packages/engine/src/__tests__/self-healing.test.ts": 11840,
   "packages/engine/src/__tests__/step-session-executor.test.ts": 3226,
-  "packages/engine/src/__tests__/triage.test.ts": 6898,
+  "packages/engine/src/__tests__/triage.test.ts": 7002,
   "packages/engine/src/agent-heartbeat.ts": 5422,
-  "packages/engine/src/agent-tools.ts": 5384,
-  "packages/engine/src/executor.ts": 19940,
+  "packages/engine/src/agent-tools.ts": 5433,
+  "packages/engine/src/executor.ts": 20076,
   "packages/engine/src/merger-ai.ts": 2401,
   "packages/engine/src/merger.ts": 11239,
-  "packages/engine/src/pi.ts": 2815,
+  "packages/engine/src/pi.ts": 2836,
   "packages/engine/src/project-engine.ts": 5410,
   "packages/engine/src/scheduler.ts": 3584,
-  "packages/engine/src/self-healing.ts": 12790,
-  "packages/engine/src/triage.ts": 3391,
+  "packages/engine/src/self-healing.ts": 12888,
+  "packages/engine/src/triage.ts": 3479,
   "packages/pi-claude-cli/src/__tests__/provider.test.ts": 2028,
   "plugins/fusion-plugin-roadmap/src/dashboard/RoadmapsView.tsx": 2583,
   "scripts/__tests__/test-changed.test.mjs": 2134
